Added JSON Web Token, Protected Routes using JWT

+++ backend/controllers/goalsController.js
@@ -3,6 +3,9 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
+//All routes are protected will get and display only specific user goals
 
 //@desc Get Goal
 //@route GET /api/Goal
@@ -10,7 +13,7 @@ const Goal = require('../models/goalModel')
 const getGoals = asyncHandler(async (req, res) => {
     //get goals through our mongoDB returns all of the goals
 
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -27,7 +30,8 @@ const createGoal = asyncHandler(async (req, res) => {
     }
     const goal = await Goal.create({
 
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
 
     })
 
@@ -45,6 +49,19 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+    //get the user
+    const user = await User.findById(req.user.id)
+    //Check for  user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make sure user matches the record that is trying to be deleted
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     //New set to true will tell mongoose to create it if the goal does not exist
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.
         body, {
@@ -65,6 +82,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal does not exist')
     }
+    //get the user
+    const user = await User.findById(req.user.id)
+    //Check for  user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make sure user matches the record that is trying to be deleted
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id, req.body)
